Added DeepLinkHelper class
